+++ aegisub/audio_provider_ram.h
@@ -55,5 +55,5 @@ public:
 	RAMAudioProvider(AudioProvider *source);
 	~RAMAudioProvider();
 
-	void GetAudio(void *buf, __int64 start, __int64 count);
+	void GetAudio(void *buf, long long start, long long count);
 };
