+++ aegisub/utils.cpp
@@ -47,7 +47,7 @@
 #ifndef __LINUX__
 //////////////////////////
 // Absolute of 64 bit int
-__int64 abs64(__int64 input) {
+long long abs64(long long input) {
 	if (input < 0) return -input;
 	return input;
 }
